osmo-bts-trx/l1_if.c: cosmetic: fix typo in debug msg
There is a big difference between both L1 and L2 frames. In the
first case, a TDMA frame is assumed, while a L2 frame is the
result of decoding a few bursts (e.g. 4 for xCCH) arrived
on a few TDMA frames.

Anyway, the purpose of l1if_process_meas_res() is to handle
measurements, not frames.

Change-Id: Ic4362c2869e658c2f42664b4009a30e777f8048c