Merging parts of commit 653e974fec00414ba85baf258ccc46ea778a53bd from branch nrw/litecell15 Original Author: Minh-Quang Nguyen <minh-quang.nguyen@nutaq.com>  LC15: Implementation of LC15 specific features
Made some modifications to this commit
based on d8cd756da48bcff9f9caf9e64fc0af047a165b60
 Get rid of 'struct gsm_bts_role_bts'

Only compile tested.
This is just to get a start on this branch.
I have not yet looked at all other commits in the
nuran branch that may be related.

This one however, does seem to be important
as it implements a larger cell radius than
is currently possible with osmo master.

Change-Id: I79416faaa3ba328c9c2dabcd695a1b880fe666da